don't bias short sequences in recurrent GAN
Before now, the generator output a "termination probability". Thus, it was much easier for the network to output short sequences than long ones. As a result, you'd get networks that collapsed to always generating sentences like "A.", "N.", ".", or "?", and the discriminator would quickly learn to identify them.

By new approach is to use a softmax over all of the possible lengths. This way, there is no inherent bias towards shorter sequences.