text gan: take turns training generator and discriminator
This is basically what I was doing manually anyway. First I'd train the generator a bit, then switch the discriminator to have an advantage and learn to reject the generator's samples, etc. Now, that process can happen naturally without my intervention. The strategy is also nice because it gives the generator a chance to discretize a bit before giving the discriminator another turn (i.e. the generator can focus on a single character per timestep).